move code from app.rb and permissioner.rb to permissions_utils.rb
